Split GUI folder browser actions
