+++ src/gui_app.rs
@@ -27,6 +27,8 @@ mod context_menu_actions;
 mod drag_drop_actions;
 mod file_actions;
 mod folder_browser;
+mod folder_browser_actions;
+mod folder_browser_rename_actions;
 mod folder_scan_actions;
 mod playback;
 mod sample_browser_view;
@@ -275,244 +277,6 @@ impl GuiAppState {
         .map_err(|err| err.to_string())
     }
 
-    fn resize_folder_browser(&mut self, message: DragHandleMessage) {
-        let started_at = Instant::now();
-        let outcome = match message {
-            DragHandleMessage::Started { .. } => "started",
-            DragHandleMessage::Moved { .. } => "moved",
-            DragHandleMessage::Ended { .. } => "ended",
-        };
-        match message {
-            DragHandleMessage::Started { position } => {
-                self.folder_resize = Some(FolderResize {
-                    start_x: position.x,
-                    start_width: self.folder_width,
-                });
-            }
-            DragHandleMessage::Moved { position } | DragHandleMessage::Ended { position } => {
-                let resize = self.folder_resize.unwrap_or(FolderResize {
-                    start_x: position.x,
-                    start_width: self.folder_width,
-                });
-                self.folder_width = (resize.start_width + position.x - resize.start_x)
-                    .clamp(MIN_FOLDER_WIDTH, MAX_FOLDER_WIDTH);
-                if matches!(message, DragHandleMessage::Ended { .. }) {
-                    self.folder_resize = None;
-                }
-            }
-        }
-        emit_gui_action(
-            "layout.resize_folder_browser",
-            Some("folder_browser"),
-            None,
-            outcome,
-            started_at,
-            None,
-        );
-    }
-
-    fn apply_folder_browser_message(
-        &mut self,
-        message: FolderBrowserMessage,
-        context: &mut ui::UpdateContext<GuiMessage>,
-    ) {
-        match message {
-            FolderBrowserMessage::AddSource => self.add_source_from_dialog(context),
-            FolderBrowserMessage::SelectSource(id) => {
-                self.context_menu = None;
-                self.select_source(id, context);
-            }
-            FolderBrowserMessage::OpenSourceContextMenu(source_id, position) => {
-                self.open_source_context_menu(source_id, position);
-            }
-            FolderBrowserMessage::BeginRenameSelected => self.begin_folder_browser_rename(context),
-            FolderBrowserMessage::BeginCreateSubfolder => {
-                self.begin_folder_browser_subfolder_creation(context);
-            }
-            FolderBrowserMessage::RenameInput(message) => {
-                self.apply_folder_browser_rename_input(message);
-            }
-            FolderBrowserMessage::DropOnFolder(folder_id) => {
-                self.context_menu = None;
-                self.drop_browser_drag_on_folder(folder_id, context);
-            }
-            FolderBrowserMessage::OpenFolderContextMenu(folder_id, position) => {
-                self.open_folder_context_menu(folder_id, position);
-            }
-            FolderBrowserMessage::DragFolder(folder_id, drag) => {
-                self.context_menu = None;
-                self.drag_folder(folder_id, drag, context);
-            }
-            message => self.folder_browser.apply_message(message),
-        }
-    }
-
-    fn begin_folder_browser_rename(&mut self, context: &mut ui::UpdateContext<GuiMessage>) {
-        let started_at = Instant::now();
-        let target = self.folder_browser.selected_rename_target();
-        if logging::debug_logging_enabled() {
-            tracing::debug!(
-                target: logging::ACTION_EVENT_TARGET,
-                event = "action_detail",
-                action = "folder_browser.rename.begin",
-                pane = "folder_browser",
-                target_kind = target.kind,
-                target_label = target.label,
-                is_source_root = target.is_source_root,
-                "Folder browser rename requested"
-            );
-        }
-        let renaming_file = self.folder_browser.selected_file_id().is_some();
-        match self.folder_browser.begin_rename_selected() {
-            Ok(Some(input_id)) => {
-                self.sample_status = if renaming_file {
-                    String::from("Renaming selected file")
-                } else {
-                    String::from("Renaming selected folder")
-                };
-                context.after(
-                    Duration::from_millis(1),
-                    GuiMessage::FocusRenameInput(input_id),
-                );
-                emit_gui_action(
-                    "folder_browser.rename.begin",
-                    Some("folder_browser"),
-                    Some(target.kind),
-                    "success",
-                    started_at,
-                    None,
-                );
-            }
-            Ok(None) => {
-                self.sample_status = String::from("Select a folder to rename");
-                emit_gui_action(
-                    "folder_browser.rename.begin",
-                    Some("folder_browser"),
-                    None,
-                    "short_circuit",
-                    started_at,
-                    Some("nothing_selected"),
-                );
-            }
-            Err(error) => {
-                self.sample_status = error;
-                emit_gui_action(
-                    "folder_browser.rename.begin",
-                    Some("folder_browser"),
-                    None,
-                    "error",
-                    started_at,
-                    Some("rename_begin_failed"),
-                );
-            }
-        }
-    }
-
-    fn begin_folder_browser_subfolder_creation(
-        &mut self,
-        context: &mut ui::UpdateContext<GuiMessage>,
-    ) {
-        let started_at = Instant::now();
-        match self.folder_browser.begin_create_subfolder() {
-            Ok(Some(input_id)) => {
-                self.sample_status = String::from("Creating new folder");
-                context.after(
-                    Duration::from_millis(1),
-                    GuiMessage::FocusRenameInput(input_id),
-                );
-                emit_gui_action(
-                    "folder_browser.folder.create_begin",
-                    Some("folder_browser"),
-                    Some("folder"),
-                    "success",
-                    started_at,
-                    None,
-                );
-            }
-            Ok(None) => {
-                self.sample_status = String::from("Select a folder to add a subfolder");
-                emit_gui_action(
-                    "folder_browser.folder.create_begin",
-                    Some("folder_browser"),
-                    None,
-                    "short_circuit",
-                    started_at,
-                    Some("nothing_selected"),
-                );
-            }
-            Err(error) => {
-                self.sample_status = error;
-                emit_gui_action(
-                    "folder_browser.folder.create_begin",
-                    Some("folder_browser"),
-                    None,
-                    "error",
-                    started_at,
-                    Some("create_begin_failed"),
-                );
-            }
-        }
-    }
-
-    fn apply_folder_browser_rename_input(&mut self, message: radiant::widgets::TextInputMessage) {
-        let started_at = Instant::now();
-        let input_action = rename_input_action(&message);
-        if let Some(status) = self.folder_browser.apply_rename_input(message) {
-            self.sample_status = status;
-        }
-        if let Some(action) = input_action {
-            emit_gui_action(
-                action,
-                Some("folder_browser"),
-                None,
-                "applied",
-                started_at,
-                None,
-            );
-        }
-    }
-
-    fn navigate_browser(
-        &mut self,
-        delta: i32,
-        extend: bool,
-        context: &mut ui::UpdateContext<GuiMessage>,
-    ) {
-        let started_at = Instant::now();
-        let direction = if delta < 0 { "previous" } else { "next" };
-        let Some(path) = self.folder_browser.navigate_vertical(delta, extend) else {
-            emit_gui_action(
-                "folder_browser.navigate",
-                Some("browser"),
-                Some(direction),
-                "edge",
-                started_at,
-                None,
-            );
-            return;
-        };
-
-        if let Some(index) = self.folder_browser.selected_audio_file_index() {
-            context.scroll_fixed_row_into_view(
-                SAMPLE_BROWSER_LIST_ID,
-                index,
-                SAMPLE_BROWSER_ROW_HEIGHT,
-                SAMPLE_BROWSER_EDGE_CONTEXT_ROWS,
-                SAMPLE_BROWSER_EDGE_CONTEXT_ROWS,
-                delta,
-            );
-        }
-        emit_gui_action(
-            "folder_browser.navigate",
-            Some("browser"),
-            Some(direction),
-            "selected",
-            started_at,
-            None,
-        );
-        self.select_sample(path, context);
-    }
-
     fn advance_frame(&mut self) {
         self.waveform.apply_interaction(WaveformInteraction::Frame);
         self.refresh_playback_progress();
@@ -543,62 +307,6 @@ impl GuiAppState {
         self.audio_sample_rate_dropdown_open = !self.audio_sample_rate_dropdown_open;
         self.audio_backend_dropdown_open = false;
         self.audio_output_dropdown_open = false;
-    }
-
-    fn focus_rename_input(&mut self, input_id: u64, context: &mut ui::UpdateContext<GuiMessage>) {
-        let started_at = Instant::now();
-        context.focus(input_id);
-        emit_gui_action(
-            "folder_browser.rename.focus_input",
-            Some("folder_browser"),
-            None,
-            "success",
-            started_at,
-            None,
-        );
-    }
-
-    fn select_all_samples(&mut self) {
-        let started_at = Instant::now();
-        let count = self.folder_browser.select_all_audio_files();
-        self.sample_status = format!(
-            "Selected {count} sample{}",
-            if count == 1 { "" } else { "s" }
-        );
-        emit_gui_action(
-            "browser.select_all_samples",
-            Some("browser"),
-            None,
-            "success",
-            started_at,
-            None,
-        );
-    }
-
-    fn collapse_selected_folder(&mut self) {
-        let started_at = Instant::now();
-        self.folder_browser.collapse_selected_folder();
-        emit_gui_action(
-            "folder_browser.collapse_selected",
-            Some("folder_browser"),
-            None,
-            "success",
-            started_at,
-            None,
-        );
-    }
-
-    fn expand_selected_folder(&mut self) {
-        let started_at = Instant::now();
-        self.folder_browser.expand_selected_folder();
-        emit_gui_action(
-            "folder_browser.expand_selected",
-            Some("folder_browser"),
-            None,
-            "success",
-            started_at,
-            None,
-        );
     }
 
     fn apply_waveform_message(&mut self, message: WaveformInteraction) {
@@ -1287,15 +995,6 @@ fn emit_gui_action(
         elapsed: started_at.elapsed(),
         error,
     });
-}
-
-fn rename_input_action(message: &radiant::widgets::TextInputMessage) -> Option<&'static str> {
-    match message {
-        radiant::widgets::TextInputMessage::Submitted { .. } => {
-            Some("folder_browser.rename.submit")
-        }
-        _ => None,
-    }
 }
 
 fn waveform_interaction_action(interaction: &WaveformInteraction) -> Option<&'static str> {
